+++ src/app/app.module.ts
@@ -142,6 +142,7 @@ import { MaskLegacyDemoComponent } from './mask/mask-legacy.demo';
 import { MenuButtonDemoComponent } from './menu-button/menu-button.demo';
 import { MessageDemoComponent } from './message/message.demo';
 import { ModalDialogDemoModule } from './modal-dialog/modal-dialog.demo.module';
+import { PagerStandaloneDemoComponent } from './pager/pager-standalone.demo';
 import { PieDemoComponent } from './pie/pie.demo';
 import { PopDownDemoComponent } from './popdown/popdown.demo';
 import { PopupMenuDemoComponent } from './popupmenu/popupmenu.demo';
@@ -320,6 +321,7 @@ import { WizardDemoValidationRulesPageComponent } from './wizard/wizard-validati
     MaskLegacyDemoComponent,
     MenuButtonDemoComponent,
     MessageDemoComponent,
+    PagerStandaloneDemoComponent,
     PieDemoComponent,
     PopDownDemoComponent,
     PopupMenuDemoComponent,